adds basic name sorting tests
